YC01: 对于办理进路后又扳动进路中的道岔导致CI不断反馈办理进路等待的信息的处理
如果接收到CI的命令反馈即设置status为4，那么当CI反馈等待信息时就会导致后续的同一命令反馈被丢弃；
所以只对命令反馈的有效性做判别，有效的命令反馈都转发给客户端处理。

	modified:   pom.xml
	modified:   src/main/java/com/byd/ats/entity/CLient2serJsonCommand.java
	modified:   src/main/java/com/byd/ats/rabbitmq/Tut5Receiver.java
	modified:   src/main/java/com/byd/ats/service/Client2serJsonCommandRepository.java

+++ src/main/java/com/byd/ats/entity/CLient2serJsonCommand.java
@@ -23,7 +23,7 @@ public class CLient2serJsonCommand {
 	private int ret;
 	private Date createtime;
 	private int status = 0; // 0:无效; 1:客户端发送命令至运行控制模块；2：运行控制模块下发命令给CU；3：收到CU的命令反馈；4：将命令反馈发送给客户端
-	private short magic; // 魔数，用于CI命令反馈信息的匹配
+	private int magic; // 魔数(1~65535)，用于CI命令反馈信息的匹配。由于CI只处理2个字节，需将值限定在1～65535范围内，而Java的short类型的值范围-32768～32767，故需改为int，不然json转换会报错。
 	private int cmd;     // 命令号
 	
 	public long getId() {
@@ -68,10 +68,10 @@ public class CLient2serJsonCommand {
 	public void setStatus(int status) {
 		this.status = status;
 	}
-	public short getMagic() {
+	public int getMagic() {
 		return magic;
 	}
-	public void setMagic(short magic) {
+	public void setMagic(int magic) {
 		this.magic = magic;
 	}
 	public int getCmd() {
